% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_funcs.R
\name{add_title_ggplot}
\alias{add_title_ggplot}
\title{Add a title to a ggplot object using cowplot.}
\usage{
add_title_ggplot(ggplot_obj, plot_title)
}
\arguments{
\item{ggplot_obj}{A ggplot object.}

\item{plot_title}{A character value for the plot title.}
}
\description{
\code{add_title_ggplot} adds a title to a ggplot object using cowplot.
}
\details{
This is useful when the ggplot object aleady has a title (e.g. gene ID) or if
it is a faceted plot (e.g. UMAP split by 'orig.ident').
}
\examples{
# load example dataset from Seurat
data("pbmc_small", package="Seurat")

add_title_ggplot(Seurat::DimPlot(pbmc_small, group.by = "RNA_snn_res.0.8", split.by = "groups"),
paste0("n=",(length(Seurat::Idents(pbmc_small)))))
}
