% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_expression.R
\name{AddAUC}
\alias{AddAUC}
\title{Calculate AUC for marker list}
\usage{
AddAUC(
  object = NULL,
  features = NULL,
  name = NULL,
  assay = NULL,
  slot = "data",
  force.recalc = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{A list of vectors of features for expression programs;
each entry should be a vector of feature names}

\item{name}{Name of the feature set (e.g. celltype) used as assay name}

\item{assay}{Name of assay to use}

\item{slot}{Name of slot to use}

\item{force.recalc}{Wheter to force recalculation}
}
\value{
SeuratObject with AUC added as 'assay'
}
\description{
Calculate the area under the curve (AUC) for each gene set for each
observation using the AUCell package
}
