% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_maxAUC}
\alias{annotate_maxAUC}
\title{Annotate clusters based on maximum AUC score}
\usage{
annotate_maxAUC(object = NULL, assay = NULL, group.by = NULL, slot = "data")
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay containing AUCs}

\item{group.by}{Vector in object's meta.data}

\item{slot}{Slot to pull data from}

\item{features}{Features to consider}
}
\description{
Annotate clusters based on maximum AUC score
}
