% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combinations}
\alias{combinations}
\title{Paste columns of a data.frame into a vector}
\usage{
combinations(x, sep = "-")
}
\arguments{
\item{x}{Data.frame}

\item{sep}{Separator put in between content of different columns}
}
\description{
Paste columns of a data.frame into a vector
}
