% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_features.R
\name{convert_names}
\alias{convert_names}
\title{Convert feature names}
\usage{
convert_names(names = NULL, dict = NULL, from = 1, to = 2)
}
\arguments{
\item{names}{Character to convert}

\item{dict}{Data frame of gene synonyms (e.g. features.tsv)}

\item{from}{Column containing the names. Can be index (integer)
or column name (character).}

\item{to}{Column containing the desired synonyms. Can be index (integer)
or column name (character).}
}
\value{
Character with synonyms
}
\description{
This is a conversion function for gene names. It takes a vector of gene names
and returns their corresponding synonyms (e.g. gene IDs) from a dictionary.
}
\examples{
# Convert human CD3D from 
convert_features("CD3D", features, 1, 2)

}
