% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_funcs.R
\name{count_cells}
\alias{count_cells}
\title{Count cells in Seurat object.}
\usage{
count_cells(seurat_object, group_by_var, subgroup_var)
}
\arguments{
\item{seurat_object}{Seurat object.}

\item{group_by_var}{A character value for grouping the cells. Must correspond
to a column in Seurat@meta.data.}

\item{subgroup_var}{Optional parameter. A character value for sub-groups
within each group. Must correspond to a column in Seurat@meta.data.}
}
\value{
Tibble with cell counts and percentages. If subgroup_var is provided,
  percentages are calculated from the sub-groups making up groups; otherwise,
  percentages are calculated from the total cell count across the groups.
}
\description{
\code{count_cells} returns a tibble with counts.
}
\examples{
# load example dataset from Seurat
data("pbmc_small", package="Seurat")

# Count group-sub-group combinations and calculate \% of sub-groups making up groups.
count_cells(pbmc_small, "groups", "RNA_snn_res.0.8")

# Count groups and calculate \% of the total that each group consists of.
count_cells(pbmc_small, "groups")

\dontrun{
count_cells(pbmc_small)
}
}
