% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_expression.R
\name{heatmap_expression}
\alias{heatmap_expression}
\title{Heatmap of gene expression}
\usage{
heatmap_expression(
  object = NULL,
  features = NULL,
  coldata = NULL,
  rowdata = NULL,
  rowdata_label = "group",
  collapse_replicates = TRUE,
  coldata_group_max = 250,
  cells = colnames(object),
  heatmap_colors = "RdBu",
  heatmap_color_dir = -1,
  annotation_colors = NULL,
  scale = TRUE,
  assay = NULL,
  slot = "data",
  limits = c(-2, 2),
  title = "Differentially expressed genes",
  show_rownames = FALSE,
  show_colnames = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Character of genes to plot}

\item{coldata}{Annotation for columns}

\item{rowdata}{Annotation for rows}

\item{rowdata_label}{Name of the row annotation}

\item{collapse_replicates}{Whether to average over groups of cells}

\item{coldata_group_max}{Maxiumum number of categories for each
column annotation}

\item{cells}{Cells to plot}

\item{heatmap_colors}{Name of color scale to use (default: "RdBu")}

\item{heatmap_color_dir}{Direction of heatmap color scale}

\item{annotation_colors}{List of color assignments for annotations}

\item{scale}{Whether to scale expression data}

\item{assay}{Which assay to take expression data from}

\item{slot}{Which slot of the assay to use ("data", "counts")}

\item{limits}{Limits of the color scale}

\item{title}{Plot title}

\item{show_rownames}{Whether to show rownames}

\item{show_colnames}{Whether to show colnames}

\item{...}{Other parameters passed to pheatmap}
}
\description{
Heatmap of gene expression
}
