% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{A Seurat object for testing.}
\format{A Seurat object:
\describe{
  \item{assays}{Contains a list of different assays. In this case, only 'RNA'.}
  \item{meta.data}{Meta data for cells.}
  ...
}}
\source{
\href{https://github.com/satijalab/seurat}{Seurat GitHub} \href{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}{Original 10x data}
}
\usage{
pbmc_small
}
\description{
A Seurat object copied from the Seurat v3.0.2 data/. It is a subsetted version of 10X Genomics' 3k PBMC dataset.
}
\seealso{
\code{\link[Seurat]{pbmc_small}}
}
\keyword{datasets}
