% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_association.R
\name{plot_association}
\alias{plot_association}
\title{Plot association between cell data}
\usage{
plot_association(object = NULL)
}
\arguments{
\item{object}{Seurat object}
}
\description{
This function creates a plot showing the association of elements of metadata
or expression of cells stored in a Seurat object.
}
