% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot low-dimensional cell embedding}
\usage{
plot_embedding(
  object = NULL,
  color = "",
  embedding = tail(names(object@reductions), 1),
  pointsize = NULL,
  dim.1 = 1,
  dim.2 = 2,
  dict = object@misc$features,
  from = 1,
  to = 2,
  brush = NULL,
  cells = NULL,
  n.cells = NULL,
  assay = "RNA",
  slot = "data"
)
}
\arguments{
\item{object}{Seurat object}

\item{color}{Gene or metadata to color cells}

\item{embedding}{Name of the embedding (e.g. 'umap')}

\item{pointsize}{Point size}

\item{dict}{Dictionary for the convert_features function}

\item{from}{From argument for convert_features}

\item{to}{To argument for convert_features}

\item{brush}{Brushed points (xmin, xmax, ymin, ymax)}

\item{cells}{Character of cell barcodes to plot}

\item{n.cells}{Number of cells to plot (randomly sampled)}

\item{assay}{Assay with expression data (e.g. RNA)}

\item{slot}{Slot of expression data (counts, data, scale.data)}
}
\description{
This function creates a plot of a cell embedding stored in a Seurat object.
Cells can be colored by gene expression or metadata.
}
\examples{
#Plot CD3D on umap embedding in Seurat object
plot_embedding(object, "CD3D", "umap")

}
