% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata.R
\name{read_h5ad}
\alias{read_h5ad}
\title{Read anndata object (.h5ad) and convert to Seurat object on disk}
\usage{
read_h5ad(file)
}
\arguments{
\item{file}{Path to h5ad object}
}
\value{
Seurat object
}
\description{
Read anndata object (.h5ad) and convert to Seurat object on disk
}
