% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata.R
\name{read_layer_h5ad}
\alias{read_layer_h5ad}
\title{Get layer from AnnData}
\usage{
read_layer_h5ad(file = NULL, name = "X")
}
\arguments{
\item{file}{Path to h5ad file of type AnnData}

\item{assay}{Assay name (default: X)}
}
\value{
Sparse matrix
}
\description{
Get layer from AnnData
}
