% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata.R
\name{read_obsm_h5ad}
\alias{read_obsm_h5ad}
\title{Read obsm from AnnData}
\usage{
read_obsm_h5ad(file)
}
\arguments{
\item{file}{Path to h5ad file of type AnnData}
}
\value{
List of matrices
}
\description{
Read obsm from AnnData
}
