% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata.R
\name{read_slot_h5ad}
\alias{read_slot_h5ad}
\title{Get slot from AnnData}
\usage{
read_slot_h5ad(file = NULL, slot = "obs")
}
\arguments{
\item{file}{Path to h5ad file of type AnnData}

\item{slot}{Slot name (default: obs)}
}
\value{
Data.frame
}
\description{
Get slot from AnnData
}
