% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{RunMNN}
\alias{RunMNN}
\title{Perform Mutual Nearest Neighbor Integration}
\usage{
RunMNN(
  object = NULL,
  batch = NULL,
  features = NULL,
  name = "mnn",
  transfer.reconstructed.assay = FALSE,
  force.recalc = FALSE,
  assay = DefaultAssay(object),
  slot = "data",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Set of features to use in MNN}

\item{name}{Name of the dimensional reduction object}

\item{transfer.reconstructed.assay}{Whether to transfer reconstructed counts
into a new assay}

\item{assay}{Assay of feature counts}

\item{slot}{Slot for feature counts}
}
\description{
Runs the fastMNN implementation of the MNN algorithm from the batchelor
package.
}
