% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{select_most_frequent_category}
\alias{select_most_frequent_category}
\title{Select maximum occuring category in vector}
\usage{
select_most_frequent_category(v)
}
\arguments{
\item{v}{Vector}
}
\description{
Select maximum occuring category in vector
}
