% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurathelpeR.R
\docType{package}
\name{seurathelpeR}
\alias{seurathelpeR}
\alias{seurathelpeR-package}
\title{seurathelpeR: A package with convenience functions for working with Seurat v3 objects.}
\description{
Simple convenience functions for working with Seurat objects.
}
\section{seurathelpeR functions}{

count_cells
add_title_ggplot
table_summary_seurat
}

