% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_groups}
\alias{summarize_groups}
\title{Summarize groups of columns in a matrix}
\usage{
summarize_groups(x = NULL, groups = NULL, FUN = sparseMatrixStats::rowMeans2)
}
\arguments{
\item{x}{A matrix-like object}

\item{groups}{A factor of group labels equal to ncol(x)}

\item{FUN}{A function to summarize columns
(default value requires sparseMatrixStats)}
}
\description{
Summarize groups of columns in a matrix
}
