% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_overlapping_rows}
\alias{summarize_overlapping_rows}
\title{Summarize rows that have overlapping coordinates}
\usage{
summarize_overlapping_rows(
  df = NULL,
  x = "x",
  y = "y",
  FUN = mean,
  breaks = 1000
)
}
\arguments{
\item{df}{Data.frame}

\item{x}{Column key}

\item{y}{Column key}

\item{FUN}{Function to summarize by}

\item{decimals}{Integer of decimals to round to}
}
\value{
Data.frame
}
\description{
Summarize rows that have overlapping coordinates
}
