% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_funcs.R
\name{table_summary_seurat}
\alias{table_summary_seurat}
\title{Function for making summary tables based on cellular data fetched using
Seurat::FetchData.}
\usage{
table_summary_seurat(fetch_dat_out, re_group)
}
\arguments{
\item{fetch_dat_out}{A data frame with cells as rows and cellular data as
columns (e.g. output from Seurat::FetchData).}

\item{re_group}{A character value for the regex to group by.}
}
\value{
A list of named tibbles, one for each metric.
}
\description{
\code{table_summary_seurat} returns a list of tibbles with summary metrics of
the metrics fetched from a Seurat object using Seurat::FetchData.
}
\examples{
# load example dataset from Seurat
data("pbmc_small", package="Seurat")

table_summary_seurat(Seurat::FetchData(pbmc_small, vars = c("nFeature_RNA", "nCount_RNA")), "^\\\\S+(?=_[ACGT]+$)")
}
