% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_differential-expression.R
\name{test_diffexp}
\alias{test_diffexp}
\title{Test differential expression}
\usage{
test_diffexp(
  object = NULL,
  groups = NULL,
  block = NULL,
  min.lfc = 0.25,
  assay = "RNA",
  slot = "data",
  dict = object@misc$features,
  from = 1,
  to = 2
)
}
\arguments{
\item{object}{Seurat object}

\item{groups}{Vector of group assignments}

\item{block}{Factor of group assignments to block}

\item{min.lfc}{Numeric scalar indicating the minimum difference in means to
calculate a p.value}

\item{assay}{Assay with expression data (e.g. RNA)}

\item{slot}{Slot of expression data (counts, data, scale.data)}
}
\description{
This function performs differential expression analysis on a Seurat object
}
