% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{to_sce}
\alias{to_sce}
\title{Convert from Seurat to SingleCellExperiment}
\usage{
to_sce(object = NULL, assay = NULL)
}
\arguments{
\item{object}{Object of class 'Seurat'}

\item{assay}{Name of assay object to use for the main SingleCellExperiment
(all other assays will be stored as altExps, defaults to first assay in list)}
}
\value{
Object of class 'SingleCellExperiment'
}
\description{
This is a conversion function between R objects from class 'Seurat' to
'SingleCellExperiment' to increase interoperability.
}
\examples{
sce <- to_sce(ds)

}
