% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_combinations}
\alias{unique_combinations}
\title{Create unique combinations from a data.frame of factors}
\usage{
unique_combinations(x, sep = "-")
}
\arguments{
\item{x}{Data.frame of factors for which to create unique combinations.
Alternatively, a list of factor levels.}
}
\description{
Create unique combinations from a data.frame of factors
}
