% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_expression.R
\name{violin_expression}
\alias{violin_expression}
\title{Create violin plot of gene expression}
\usage{
violin_expression(
  object = NULL,
  features = NULL,
  coldata = NULL,
  max.features = 50,
  assay = NULL,
  slot = "data",
  cells = colnames(object)
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Vector of features, defaults to whole assay}

\item{assay}{Name of assay to use}

\item{slot}{Name of slot to use}

\item{max_features}{Maximum number of features to plot}
}
\description{
Create violin plot of gene expression
}
