% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_rowMax}
\alias{which_rowMax}
\title{Extract the column names of row maxima}
\usage{
which_rowMax(x)
}
\arguments{
\item{x}{Matrix containing numeric values}
}
\description{
Extract the column names of row maxima
}
