% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata.R
\name{write_h5ad}
\alias{write_h5ad}
\title{Convert Seurat to h5ad}
\usage{
write_h5ad(object = NULL, file = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{file}{Path to h5ad object (e.g. "~/Downloads/pbmc3k.h5ad")}
}
\description{
Save on disk Seurat object as anndata object (.h5ad)
}
