% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_expression.R
\name{heatmap_expression}
\alias{heatmap_expression}
\title{Heatmap of gene expression}
\usage{
heatmap_expression(
  object = NULL,
  features = NULL,
  coldata = NULL,
  rowdata = NULL,
  rowdata_label = "group",
  collapse_replicates = TRUE,
  colors = NULL,
  scale = TRUE,
  assay = NULL,
  slot = "data",
  limits = c(-2, 2),
  title = "Differentially expressed genes",
  show_rownames = FALSE,
  show_colnames = FALSE,
  fontsize = 10,
  fontsize_row = fontsize,
  fontsize_col = fontsize
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Character of genes to plot}

\item{coldata}{Annotation for columns}

\item{rowdata}{Annotation for rows}

\item{rowdata_label}{Name of the row annotation}

\item{collapse_replicates}{Whether to average over groups of cells}

\item{limits}{Limits of the color scale}

\item{title}{Plot title}
}
\description{
Heatmap of gene expression
}
