% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot low-dimensional cell embedding}
\usage{
plot_embedding(
  object = NULL,
  color = "",
  label = FALSE,
  label.size = 4,
  embedding = tail(names(object@reductions), 1),
  pt.size = NULL,
  dim.1 = 1,
  dim.2 = 2,
  brush = NULL,
  cells = NULL,
  n.cells = NULL,
  assay = "RNA",
  slot = "data",
  dict = object@misc$features,
  from = 1,
  to = 2,
  alpha = 1,
  shape = 20,
  color.transform = "",
  legend.position = "right",
  legend.rows = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{color}{Gene or metadata to color cells}

\item{label}{Type of group labels (text or label)}

\item{embedding}{Name of the embedding (e.g. 'umap')}

\item{brush}{Brushed points (xmin, xmax, ymin, ymax)}

\item{cells}{Character of cell barcodes to plot}

\item{n.cells}{Number of cells to plot (randomly sampled)}

\item{assay}{Assay with expression data (e.g. RNA)}

\item{slot}{Slot of expression data (counts, data, scale.data)}

\item{dict}{Dictionary for the convert_features function}

\item{from}{From argument for convert_features}

\item{to}{To argument for convert_features}

\item{alpha}{Transparency of points (0-1)}

\item{color.transform}{Transformation of values used for color.
Options include log, log2, log1p, log10}

\item{legend.position}{Position of the color legend.}

\item{pointsize}{Point size}
}
\description{
This function creates a plot of a cell embedding stored in a Seurat object.
Cells can be colored by gene expression or metadata.
}
