% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{plot_markers_embedding}
\alias{plot_markers_embedding}
\title{Plot cell type markers on embedding}
\usage{
plot_markers_embedding(
  object,
  markers = NULL,
  embedding = tail(names(object@reductions), 1),
  pt.aggr = TRUE,
  pt.aggr.breaks = 300,
  pt.size = 0.5,
  pt.stroke = 0.1,
  pt.shape = 16,
  nrow = floor(sqrt(length(markers))),
  assay = NULL,
  slot = "data",
  cells = NULL,
  scale = TRUE,
  markers.max = 100,
  pl.title = NULL,
  col.title = NULL
)
}
\arguments{
\item{object}{SeuratObject}

\item{markers}{Character vector of marker genes}

\item{embedding}{Embedding name (slot in reductions)}

\item{pt.size}{Point size}

\item{nrow}{Number of rows}

\item{cells}{Character vector of cells to plot}
}
\value{
plot
}
\description{
Plot cell type markers on embedding
}
