% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{plot_markers_embedding}
\alias{plot_markers_embedding}
\title{Plot cell type markers on embedding}
\usage{
plot_markers_embedding(
  object,
  markers = NULL,
  embedding = tail(names(object@reductions), 1),
  nrow = 4,
  pt.size = 1
)
}
\arguments{
\item{object}{SeuratObject}

\item{markers}{Character vector of marker genes}

\item{embedding}{Embedding name (slot in reductions)}

\item{nrow}{Number of rows}

\item{pt.size}{Point size}
}
\value{
plot
}
\description{
Plot cell type markers on embedding
}
