% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_qc_metrics}
\alias{plot_qc_metrics}
\title{Plot QC metrics}
\usage{
plot_qc_metrics(
  object,
  libsize = "libsize",
  metrics = c("ngenes", "percent.mt"),
  quality = NULL,
  color_by = NULL
)
}
\arguments{
\item{object}{SeuratObject}
}
\value{
plot
}
\description{
Plot QC metrics
}
